requirements.txt
altair==5.5.0
asgiref==3.9.1
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
django==5.2.4
djangorestframework==3.16.0
gitdb==4.0.12
gitpython==3.1.44
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markupsafe==3.0.2
mysqlclient==2.2.7
narwhals==1.47.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pip==24.2
pluggy==1.6.0
protobuf==6.31.1
pyarrow==20.0.0
pydeck==0.9.1
pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
sqlparse==0.5.3
streamlit==1.47.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing-extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
django-cors-headers==4.3.1